Pull functionality up from sql_datasource.py to interfaces.py
